+++ data/player.go
@@ -352,6 +352,23 @@ func GetPlayersScore(legID int) (map[int]*models.Player2Leg, error) {
 				scores[visit.PlayerID].CurrentScore += score
 			}
 		}
+	} else if m.MatchType.ID == models.FOURTWENTY {
+		visits, err := GetLegVisits(legID)
+		if err != nil {
+			return nil, err
+		}
+		for _, player := range scores {
+			player.CurrentScore = 420
+		}
+
+		round := 1
+		for i, visit := range visits {
+			if i > 0 && i%len(players) == 0 {
+				round++
+			}
+			score := visit.Calculate420Score(round - 1)
+			scores[visit.PlayerID].CurrentScore -= score
+		}
 	}
 
 	return scores, nil