+++ models/match.go
@@ -28,22 +28,46 @@ const (
 	TICTACTOE = 9
 	// BERMUDATRIANGLE constant representing type 10
 	BERMUDATRIANGLE = 10
+	// FOURTWENTY constant representing type 11
+	FOURTWENTY = 11
 )
 
 // TargetsBermudaTriangle contains the target for each round of Bermuda Triangle
-var TargetsBermudaTriangle = [13]Target{Target{value: 12, multipliers: []int64{1, 2, 3}},
-	Target{value: 13, multipliers: []int64{1, 2, 3}},
-	Target{value: 14, multipliers: []int64{1, 2, 3}},
-	Target{value: -1, multipliers: []int64{2}},
-	Target{value: 15, multipliers: []int64{1, 2, 3}},
-	Target{value: 16, multipliers: []int64{1, 2, 3}},
-	Target{value: 17, multipliers: []int64{1, 2, 3}},
-	Target{value: -1, multipliers: []int64{3}},
-	Target{value: 18, multipliers: []int64{1, 2, 3}},
-	Target{value: 19, multipliers: []int64{1, 2, 3}},
-	Target{value: 20, multipliers: []int64{1, 2, 3}},
-	Target{value: 25, multipliers: []int64{1, 2}, score: 25},
-	Target{value: 25, multipliers: []int64{2}}}
+var TargetsBermudaTriangle = [13]Target{Target{Value: 12, multipliers: []int64{1, 2, 3}},
+	Target{Value: 13, multipliers: []int64{1, 2, 3}},
+	Target{Value: 14, multipliers: []int64{1, 2, 3}},
+	Target{Value: -1, multipliers: []int64{2}},
+	Target{Value: 15, multipliers: []int64{1, 2, 3}},
+	Target{Value: 16, multipliers: []int64{1, 2, 3}},
+	Target{Value: 17, multipliers: []int64{1, 2, 3}},
+	Target{Value: -1, multipliers: []int64{3}},
+	Target{Value: 18, multipliers: []int64{1, 2, 3}},
+	Target{Value: 19, multipliers: []int64{1, 2, 3}},
+	Target{Value: 20, multipliers: []int64{1, 2, 3}},
+	Target{Value: 25, multipliers: []int64{1, 2}, score: 25},
+	Target{Value: 25, multipliers: []int64{2}}}
+
+// Targets420 contains the target for each round of 420
+var Targets420 = [20]Target{Target{Value: 1, multipliers: []int64{2}},
+	{Value: 18, multipliers: []int64{2}},
+	{Value: 4, multipliers: []int64{2}},
+	{Value: 13, multipliers: []int64{2}},
+	{Value: 6, multipliers: []int64{2}},
+	{Value: 10, multipliers: []int64{2}},
+	{Value: 15, multipliers: []int64{2}},
+	{Value: 2, multipliers: []int64{2}},
+	{Value: 17, multipliers: []int64{2}},
+	{Value: 3, multipliers: []int64{2}},
+	{Value: 19, multipliers: []int64{2}},
+	{Value: 7, multipliers: []int64{2}},
+	{Value: 16, multipliers: []int64{2}},
+	{Value: 8, multipliers: []int64{2}},
+	{Value: 11, multipliers: []int64{2}},
+	{Value: 14, multipliers: []int64{2}},
+	{Value: 9, multipliers: []int64{2}},
+	{Value: 12, multipliers: []int64{2}},
+	{Value: 5, multipliers: []int64{2}},
+	{Value: 20, multipliers: []int64{2}}}
 
 // Match struct used for storing matches
 type Match struct {
@@ -199,7 +223,7 @@ type MatchMetadata struct {
 
 // Target contains information about value and multipler required to hit for a given round
 type Target struct {
-	value       int
+	Value       int
 	multipliers []int64
 	score       int
 }